C65: FDC/D81 access cleanup
