--- targets/c65/c65_d81_image.h
@@ -1,25 +1,0 @@
-/* Test-case for a very simple, inaccurate, work-in-progress Commodore 65 emulator.
-   Copyright (C)2016,2018 LGB (Gábor Lénárt) <lgblgblgb@gmail.com>
-
-This program is free software; you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation; either version 2 of the License, or
-(at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program; if not, write to the Free Software
-Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA */
-
-#ifndef __XEMU_C65_D81_IMAGE_H_INCLUDED
-#define __XEMU_C65_D81_IMAGE_H_INCLUDED
-
-extern void c65_d81_init ( const char *dfn, int force_ro );
-
-extern Uint8 disk_cache[512];	// internal memory of the F011 disk controller
-
-#endif
