+++ xemu/f011_core.h
@@ -19,7 +19,9 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA */
 #ifndef __XEMU_F011_CORE_H_INCLUDED
 #define __XEMU_F011_CORE_H_INCLUDED
 
+#ifndef D81_SIZE
 #define D81_SIZE 819200
+#endif
 
 extern void  fdc_write_reg ( int addr, Uint8 data );
 extern Uint8 fdc_read_reg  ( int addr );
